+++ app/constants/translations/de/navigate.js
@@ -1,14 +1,21 @@
 export default {
-  home: 'Heim',
-  setting: 'Einstellungen',
-  description:
-    'Das Cushing-Syndrom tritt auf, wenn der Körper ' +
-    'mit der Zeit zu viel des Hormons Cortisol hat. ' +
-    'Es kann das Ergebnis der oralen Einnahme von ' +
-    'Kortikosteroiden oder der Körperproduktion ' +
-    'von zu viel Cortisol sein.',
+  home: {
+    title: 'Heim',
+  },
+  setting: {
+    title: 'Einstellung',
+  },
+  description: {
+    title: 'Bezeichnung',
+    description:
+      'Das Cushing-Syndrom tritt auf, wenn der Körper ' +
+      'mit der Zeit zu viel des Hormons Cortisol hat. ' +
+      'Es kann das Ergebnis der oralen Einnahme von ' +
+      'Kortikosteroiden oder der Körperproduktion ' +
+      'von zu viel Cortisol sein.',
+  },
   symptom: {
-    title: 'Häufige Symptome',
-    description: 'Häufige Symptomes',
+    title: 'Symptome',
+    description: 'Symptome',
   },
 };