+++ app/constants/translations/en/navigate.js
@@ -1,14 +1,21 @@
 export default {
-  home: 'Home',
-  setting: 'Settings',
-  description:
-    "Cushing's syndrome occurs when the body " +
-    'has too much of the hormone cortisol over ' +
-    'time. It can be the result of taking ' +
-    'corticosteroids by mouth or the body making ' +
-    'too much cortisol.',
+  home: {
+    title: 'Home',
+  },
+  setting: {
+    title: 'Setting',
+  },
+  description: {
+    title: 'Description',
+    description:
+      "Cushing's syndrome occurs when the body " +
+      'has too much of the hormone cortisol over ' +
+      'time. It can be the result of taking ' +
+      'corticosteroids by mouth or the body making ' +
+      'too much cortisol.',
+  },
   symptom: {
-    title: 'Common symptoms',
-    description: 'Common symptoms',
+    title: 'Symptoms',
+    description: 'Symptoms',
   },
 };