+++ app/constants/translations/es/navigate.js
@@ -1,14 +1,21 @@
 export default {
-  home: 'Inicio',
-  setting: 'Configuraciones',
-  description:
-    'El síndrome de Cushing aparece cuando ' +
-    'el cuerpo tiene demasiada hormona cortisol a lo largo ' +
-    'del tiempo. Puede ser el resultado de tomar ' +
-    'corticosteroides por vía oral o de que el' +
-    ' cuerpo produzca demasiado cortisol.',
+  home: {
+    title: 'Inicio',
+  },
+  setting: {
+    title: 'Configuración',
+  },
+  description: {
+    title: 'Descripción',
+    description:
+      'El síndrome de Cushing aparece cuando ' +
+      'el cuerpo tiene demasiada hormona cortisol a lo largo ' +
+      'del tiempo. Puede ser el resultado de tomar ' +
+      'corticosteroides por vía oral o de que el' +
+      ' cuerpo produzca demasiado cortisol.',
+  },
   symptom: {
-    title: 'Síntomas comunes',
-    description: 'Síntomas comunes',
+    title: 'Síntomas',
+    description: 'Síntomas',
   },
 };