+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.4.1
 
 - [fix] Description screen.
+- [refactor] Screens.
 
 # v0.4.0